+++ source/extensions/omni.flux.validator.plugin.selector.usd/omni/flux/validator/plugin/selector/usd/extension.py
@@ -20,6 +20,7 @@ import carb.settings
 import omni.ext
 from omni.flux.validator.factory import get_instance as _get_factory_instance
 
+from .all_lights import AllLights as _AllLights
 from .all_materials import AllMaterials as _AllMaterials
 from .all_meshes import AllMeshes as _AllMeshes
 from .all_prims import AllPrims as _AllPrims
@@ -35,11 +36,11 @@ class FluxValidatorPluginSelectorUSDExtension(omni.ext.IExt):
     def on_startup(self, ext_id):
         carb.log_info("[omni.flux.validator.plugin.selector.usd] Startup")
         _get_factory_instance().register_plugins(
-            [_AllPrims, _AllMeshes, _AllMaterials, _AllShaders, _AllTextures, _Nothing, _RootPrims]
+            [_AllLights, _AllPrims, _AllMeshes, _AllMaterials, _AllShaders, _AllTextures, _Nothing, _RootPrims]
         )
 
     def on_shutdown(self):
         carb.log_info("[omni.flux.validator.plugin.selector.usd] Shutdown")
         _get_factory_instance().unregister_plugins(
-            [_AllPrims, _AllMeshes, _AllMaterials, _AllShaders, _AllTextures, _Nothing, _RootPrims]
+            [_AllLights, _AllPrims, _AllMeshes, _AllMaterials, _AllShaders, _AllTextures, _Nothing, _RootPrims]
         )