+++ source/extensions/lightspeed.trex.project_wizard.core/docs/CHANGELOG.md
@@ -1,9 +1,13 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.2.15]
+### Changed
+- Update deps
+
 ## [1.2.14]
 ### Changed
-- Update dependency
+- Update deps
 
 ## [1.2.13]
 ### Changed