+++ source/extensions/omni.flux.validator.factory/docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [2.7.0]
+### Added
+- Added constants for fixed files
+
 ## [2.6.0]
 ### Changed
 - Moved `_get_schema_data_flows` function to `Base` class