+++ source/extensions/omni.flux.validator.plugin.resultor.file/docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.7.0]
+### Changed
+- `FileMetadataWritter` will now write file fixes for all output files
+
 ## [1.6.0]
 ### Added
 - Added `dataflow_to_json` resultor to save dataflow contents to a JSON file