+++ source/extensions/omni.flux.validator.plugin.selector.usd/docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.8.0]
+### Changed
+- Added the ability to select root layer prims only for all selectors using `select_from_root_layer_only`
+
 ## [1.7.4] - 2024-05-29
 ### Changed
 - File extension check now uses `get_invalid_extensions()`